added comments and moved audio to AudioEngine
